+++ packages/phpstan-rules/tests/Rules/PreventDuplicateClassMethodRule/Fixture/DifferentMethodName1.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\PHPStanRules\Tests\Rules\PreventDuplicateClassMethodRule\Fixture;
+
+class DifferentMethodName1
+{
+    public function go($value)
+    {
+        $value += 100000;
+        return $value + 1000;
+    }
+}
