+++ packages/phpstan-rules/tests/Rules/PreventDuplicateClassMethodRule/Fixture/DifferentMethodName2.php
@@ -4,15 +4,6 @@ declare(strict_types=1);
 
 namespace Symplify\PHPStanRules\Tests\Rules\PreventDuplicateClassMethodRule\Fixture;
 
-class DifferentMethodName1
-{
-    public function go($value)
-    {
-        $value += 100000;
-        return $value + 1000;
-    }
-}
-
 class DifferentMethodName2
 {
     public function sleep($yet)
