+++ packages/phpstan-rules/tests/Rules/PreventDuplicateClassMethodRule/Fixture/WithNoParameter1.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\PHPStanRules\Tests\Rules\PreventDuplicateClassMethodRule\Fixture;
+
+final class WithNoParameter1
+{
+    public function diff()
+    {
+        echo 'statement';
+        (new SmartFinder())->run('.php');
+    }
+}
