+++ packages/phpstan-rules/tests/Rules/PreventDuplicateClassMethodRule/Fixture/WithParameter1.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\PHPStanRules\Tests\Rules\PreventDuplicateClassMethodRule\Fixture;
+
+final class WithParameter1
+{
+    public function diff($param)
+    {
+        echo 'statement';
+        $param->run('.php');
+    }
+}
