+++ packages/phpstan-rules/tests/Rules/PreventDuplicateClassMethodRule/Fixture/WithParameter2.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\PHPStanRules\Tests\Rules\PreventDuplicateClassMethodRule\Fixture;
+
+final class WithParameter2
+{
+    public function diff($param)
+    {
+        echo 'statement';
+        $param->run('.php');
+    }
+}
